separation of classes
