+++ components/assistant.js
@@ -5,9 +5,6 @@ const BufferToMP3 = require("./BufferToMP3")
 const Recorder = require("./lpcm16")
 const path = require("path")
 const fs = require("fs")
-const HTMLParser = require("node-html-parser")
-const Entities = require('html-entities')
-const cheerio = require( "cheerio")
 
 class ASSISTANT {
   constructor(config, tunnel = ()=>{}) {
@@ -219,171 +216,4 @@ class ASSISTANT {
   }
 }
 
-class SCREENPARSER {
-  constructor(config, debug) {
-    this.config = config
-    if (debug == true) logGA = (...args) => { console.log("[GA] [SCREEN_PARSER]", ...args) }
-  }
-
-  parse(response, endCallback=()=>{}) {
-    if (response.screen) {
-      var uri = this.config.responseOutputURI
-      var filePath = path.resolve(__dirname, "..", uri)
-      if (!response.screen.originalContent) return
-      var str = response.screen.originalContent.toString("utf8")
-      var disableTimeoutFromScreenOutput = (str) => {
-        return str.replace(/document\.body,"display","none"/gim,(x)=>{
-          return `document.body,"display","block"`
-        })
-      }
-      str = disableTimeoutFromScreenOutput(str)
-      str = str.replace("html", 'html style="zoom:' + this.config.responseOutputZoom + '"')
-
-      var url = "/modules/MMM-GoogleAssistant/" + this.config.responseOutputCSS + "?seed=" + Date.now()
-      str = str.replace(/<style>html,body[^<]+<\/style>/gmi, `<link rel="stylesheet" href="${url}">`)
-
-      var ret = HTMLParser.parse(response.screen.originalContent)
-      var dom = ret.querySelector(".popout-content")
-      response.screen.text = dom ? dom.structuredText : null
-      response.text= dom && dom.querySelector(".show_text_content") ? dom.querySelector(".show_text_content").structuredText : null
-      response.screen = this.parseScreenLink(response.screen)
-      response.screen.photos = []
-      var photos = ret.querySelectorAll(".photo_tv_image")
-      if (photos) {
-        for (var i=0; i < photos.length; i++) {
-          response.screen.photos.push(photos[i].attributes["data-image-url"])
-        }
-      }
-
-      var contents = fs.writeFile(filePath, str, (error) => {
-        if (error) {
-          console.error("[GA] [SCREEN_PARSER] SCREENOUTPUT_CREATION_ERROR", error)
-          endCallback(error)
-        } else {
-          response.screen.path = filePath
-          response.screen.uri = uri
-          logGA("SCREEN_OUTPUT_CREATED")
-          endCallback(response)
-        }
-      })
-    }
-  }
-
-  parseScreenLink(screen) {
-    var decode = Entities.decode
-    var html = screen.originalContent
-    screen.links = []
-    var links = [
-      /data-url=\"([^\"]+)\"/gmi,
-      / (http[s]?\:\/\/[^ \)]+)[ ]?\)/gmi,
-      /\: (http[s]?\:\/\/[^ <]+)/gmi,
-    ]
-    var r = null
-    var res = []
-    for (var i = 0; i < links.length; i++) {
-      var link = links[i]
-      while ((r = link.exec(html)) !== null) {
-        res.push(decode(r[1]))
-      }
-    }
-    screen.links = res
-    logGA("[LINKS] Found:", screen.links.length)
-    return screen
-  }
-}
-
-class GoogleSearch {
-  constructor() {
-    this.defaultUserAgent = "Mozilla/5.0 (Linux x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/111.0.0.0 Safari/537.36 MMM-GoogleAssistant/"+require('../package.json').version;
-
-    this.defaultLimit = 5
-    this.defaultStart = 0
-    this.linkSelector = 'div.fP1Qef > div:nth-child(1) > a'
-
-    this.parseGoogleSearchResultUrl = (url) => {
-      if (!url) return undefined
-      if (url.charAt(0) === '/') {
-        const searchURL= new URLSearchParams(url);
-        return searchURL.get("url")
-      }
-      return url
-    }
-  }
-
-  getResults({ data }) {
-    const $ = cheerio.load(data)
-    let results = []
-
-    $(this.linkSelector).map((index, elem) => {
-      const link = this.parseGoogleSearchResultUrl(elem.attribs.href)
-      if (link.startsWith('http://www.google.com') || link.startsWith('https://www.google.com')) return
-      results.push({ link: link })
-    })
-
-    return { results }
-  }
-
-  getResponse({ query }) {
-    return new Promise((resolve, reject) => {
-      const params = new URLSearchParams({ q: query, num: this.defaultLimit, start: this.defaultStart })
-      fetch(`https://www.google.com/search?${params}`,
-        {
-          "headers": {
-            'User-Agent': this.defaultUserAgent,
-          }
-        })
-        .then(async response => {
-          const body = await response.text()
-          const status = response.status
-          return resolve({ body, status })
-        })
-        .catch((error) => reject(new Error(`[GA] [GoogleSearch] Error making web request:`, error)))
-    })
-  }
-
-  googleIt(config) {
-    return new Promise((resolve, reject) => {
-      this.getResponse(config).then(({ body, status }) => {
-        const { results } = this.getResults({ data: body })
-        if (results.length === 0 && status !== 200) reject(new Error(`[GA] [GoogleSearch] Error in response: status ${status}.`))
-        return resolve(results)
-      }).catch(reject)
-    })
-  }
-
-  search (that, text) {
-    if (!text) return
-    if (that.config.debug) logGA = (...args) => { console.log("[GA] [GoogleSearch]", ...args) }
-    var finalResult = []
-    this.googleIt({ query: text })
-      .then(results => {
-        if (results && results.length) {
-          results.forEach(link => {
-            logGA("Link:", link.link)
-            finalResult.push(link.link)
-          })
-
-          if (finalResult.length) {
-            logGA("Results:",finalResult)
-            that.sendSocketNotification("GOOGLESEARCH-RESULT", finalResult[0])
-          } else {
-            logGA("No Results found!")
-            that.sendSocketNotification("ERROR", "[GoogleSearch] No Results found!")
-          }
-        } else {
-          logGA("No Results found!")
-          that.sendSocketNotification("ERROR", "[GoogleSearch] No Results found!")
-        }
-      })
-      .catch(e => {
-        console.error("[GA] [GOOGLE_SEARCH] [ERROR] " + e)
-        that.sendSocketNotification("ERROR", "[GoogleSearch] Sorry, an error occurred!")
-      })
-  }
-};
-
-module.exports = {
-  ASSISTANT: ASSISTANT,
-  SCREENPARSER: SCREENPARSER,
-  GOOGLESEARCH: GoogleSearch
-}
+module.exports = ASSISTANT
